+++ src/Sync/Processor/Group/GroupSyncProcessor.php
@@ -238,8 +238,10 @@ class GroupSyncProcessor extends ObjectSyncProcessor implements IGroupSyncProces
 
 	/**
 	 * @inheritdoc
+	 *
+	 * @param GroupDTO $dto
 	 */
-	protected function handleDelete($ilias_id)/*: void*/ {
+	protected function handleDelete(IDataTransferObject $dto, $ilias_id)/*: void*/ {
 		$this->current_ilias_object = $ilObjGroup = $this->findILIASGroup($ilias_id);
 		if ($ilObjGroup === NULL) {
 			return;